+++ compiler-cli/src/dependencies.rs
@@ -1093,7 +1093,7 @@ async fn lookup_package(
     match provided.get(name.as_str()) {
         Some(provided_package) => Ok(provided_package.to_manifest_package(name.as_str())),
         None => {
-            let config = hexpm::Config::new();
+            let config = crate::hex::hex_config();
             let release =
                 hex::get_package_release(&name, &version, &config, &HttpClient::new()).await?;
             let build_tools = release
@@ -1183,7 +1183,7 @@ impl dependency::PackageFetcher for PackageFetcher {
         }
 
         tracing::debug!(package = package, "looking_up_hex_package");
-        let config = hexpm::Config::new();
+        let config = crate::hex::hex_config();
         let request = hexpm::repository_v2_get_package_request(package, None, &config);
         let response = self
             .runtime